refactor(sampling): consolidate kernel dispatch
Gather architecture-neutral rotation, measurement, fused-rotation, and instrument dispatch into one lowering/runtime boundary. Keep ISA-specific declarations private in one header while retaining separate scalar, AVX2, and AVX-512 implementation translation units.

This removes nine plumbing files without changing kernel selection, descriptors, or optimized hot-function code sizes.
