+++ src/clifft/sampling/kernels_avx2.cc
@@ -1,11 +1,8 @@
 // AVX2+BMI2+FMA sampling kernels. This translation unit is compiled with
 // explicit ISA flags so portable builds can select it at runtime.
 
-#include "clifft/sampling/active_measurement_simd.h"
-#include "clifft/sampling/direct_rotation_simd.h"
-#include "clifft/sampling/fused_rotation_simd.h"
 #include "clifft/sampling/indexing.h"
-#include "clifft/sampling/instrument_activation_simd.h"
+#include "clifft/sampling/simd_kernels.h"
 #include "clifft/sampling/simd_width.h"
 #include "clifft/util/numeric.h"
