+++ tests/test_sampling_kernels.cc
@@ -1,6 +1,4 @@
-#include "clifft/sampling/active_measurement_dispatch.h"
-#include "clifft/sampling/direct_rotation_dispatch.h"
-#include "clifft/sampling/instrument_activation_dispatch.h"
+#include "clifft/sampling/kernel_dispatch.h"
 #include "clifft/sampling/kernels.h"
 #include "clifft/util/runtime_isa.h"
